+++ crates/rust-client/src/store/sqlite_store/mod.rs
@@ -11,7 +11,10 @@ use alloc::{
 };
 use std::{path::PathBuf, string::ToString};
 
-use deadpool_sqlite::{Config, Hook, HookError, Pool, Runtime};
+use db_management::{
+    pool_manager::{Pool, SqlitePoolManager},
+    utils::apply_migrations,
+};
 use miden_objects::{
     Digest, Word,
     account::{Account, AccountCode, AccountHeader, AccountId},
@@ -20,12 +23,12 @@ use miden_objects::{
     note::{NoteTag, Nullifier},
     transaction::TransactionId,
 };
-use rusqlite::{Connection, types::Value, vtab::array};
+use rusqlite::{Connection, types::Value};
 use tonic::async_trait;
 
 use super::{
-    AccountRecord, AccountStatus, ChainMmrNodeFilter, InputNoteRecord, NoteFilter,
-    OutputNoteRecord, Store, TransactionFilter,
+    AccountRecord, AccountStatus, InputNoteRecord, NoteFilter, OutputNoteRecord,
+    PartialBlockchainFilter, Store, TransactionFilter,
 };
 use crate::{
     note::NoteUpdates,
@@ -36,6 +39,7 @@ use crate::{
 
 mod account;
 mod chain_data;
+mod db_management;
 mod errors;
 mod note;
 mod sync;
@@ -58,38 +62,19 @@ impl SqliteStore {
 
     /// Returns a new instance of [Store] instantiated with the specified configuration options.
     pub async fn new(database_filepath: PathBuf) -> Result<Self, StoreError> {
-        let database_exists = database_filepath.exists();
-
-        let connection_cfg = Config::new(database_filepath);
-        let pool = connection_cfg
-            .builder(Runtime::Tokio1)
-            .map_err(|err| StoreError::DatabaseError(err.to_string()))?
-            .post_create(Hook::async_fn(move |conn, _| {
-                Box::pin(async move {
-                    // Feature used to support `IN` and `NOT IN` queries. We need to load this
-                    // module for every connection we create to the DB to
-                    // support the queries we want to run
-                    conn.interact(|conn| array::load_module(conn))
-                        .await
-                        .map_err(|_| HookError::message("Loading rarray module failed"))?
-                        .map_err(|err| HookError::message(err.to_string()))?;
-
-                    Ok(())
-                })
-            }))
+        let sqlite_pool_manager = SqlitePoolManager::new(database_filepath.clone());
+        let pool = Pool::builder(sqlite_pool_manager)
             .build()
-            .map_err(|err| StoreError::DatabaseError(err.to_string()))?;
+            .map_err(|e| StoreError::DatabaseError(e.to_string()))?;
 
-        if !database_exists {
-            pool.get()
-                .await
-                .map_err(|err| StoreError::DatabaseError(err.to_string()))?
-                .interact(|conn| conn.execute_batch(include_str!("store.sql")))
-                .await
-                .map_err(|err| StoreError::DatabaseError(err.to_string()))??;
-        }
+        let conn = pool.get().await.map_err(|e| StoreError::DatabaseError(e.to_string()))?;
 
-        Ok(Self { pool })
+        let _ = conn
+            .interact(apply_migrations)
+            .await
+            .map_err(|e| StoreError::DatabaseError(e.to_string()))?;
+
+        Ok(SqliteStore { pool })
     }
 
     /// Interacts with the database by executing the provided function on a connection from the
@@ -193,7 +178,7 @@ impl Store for SqliteStore {
     async fn insert_block_header(
         &self,
         block_header: &BlockHeader,
-        chain_mmr_peaks: MmrPeaks,
+        partial_blockchain_peaks: MmrPeaks,
         has_client_notes: bool,
     ) -> Result<(), StoreError> {
         let block_header = block_header.clone();
@@ -201,7 +186,7 @@ impl Store for SqliteStore {
             SqliteStore::insert_block_header(
                 conn,
                 &block_header,
-                &chain_mmr_peaks,
+                &partial_blockchain_peaks,
                 has_client_notes,
             )
         })
@@ -223,29 +208,33 @@ impl Store for SqliteStore {
         self.interact_with_connection(SqliteStore::get_tracked_block_headers).await
     }
 
-    async fn get_chain_mmr_nodes(
+    async fn get_partial_blockchain_nodes(
         &self,
-        filter: ChainMmrNodeFilter,
+        filter: PartialBlockchainFilter,
     ) -> Result<BTreeMap<InOrderIndex, Digest>, StoreError> {
-        self.interact_with_connection(move |conn| SqliteStore::get_chain_mmr_nodes(conn, &filter))
-            .await
+        self.interact_with_connection(move |conn| {
+            SqliteStore::get_partial_blockchain_nodes(conn, &filter)
+        })
+        .await
     }
 
-    async fn insert_chain_mmr_nodes(
+    async fn insert_partial_blockchain_nodes(
         &self,
         nodes: &[(InOrderIndex, Digest)],
     ) -> Result<(), StoreError> {
         let nodes = nodes.to_vec();
-        self.interact_with_connection(move |conn| SqliteStore::insert_chain_mmr_nodes(conn, &nodes))
-            .await
+        self.interact_with_connection(move |conn| {
+            SqliteStore::insert_partial_blockchain_nodes(conn, &nodes)
+        })
+        .await
     }
 
-    async fn get_chain_mmr_peaks_by_block_num(
+    async fn get_partial_blockchain_peaks_by_block_num(
         &self,
         block_num: BlockNumber,
     ) -> Result<MmrPeaks, StoreError> {
         self.interact_with_connection(move |conn| {
-            SqliteStore::get_chain_mmr_peaks_by_block_num(conn, block_num)
+            SqliteStore::get_partial_blockchain_peaks_by_block_num(conn, block_num)
         })
         .await
     }
@@ -347,7 +336,7 @@ impl Store for SqliteStore {
 
 /// Gets a `u64` value from the database.
 ///
-/// Sqlite uses `i64` as its internal representation format, and so when retrieving
+/// `Sqlite` uses `i64` as its internal representation format, and so when retrieving
 /// we need to make sure we cast as `u64` to get the original value
 pub fn column_value_as_u64<I: rusqlite::RowIndex>(
     row: &rusqlite::Row<'_>,
@@ -363,8 +352,8 @@ pub fn column_value_as_u64<I: rusqlite::RowIndex>(
 
 /// Converts a `u64` into a [Value].
 ///
-/// Sqlite uses `i64` as its internal representation format. Note that the `as` operator performs a
-/// lossless conversion from `u64` to `i64`.
+/// `Sqlite` uses `i64` as its internal representation format. Note that the `as` operator performs
+/// a lossless conversion from `u64` to `i64`.
 pub fn u64_to_value(v: u64) -> Value {
     #[allow(
         clippy::cast_possible_wrap,