+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 * [BREAKING] Updated keystore to accept arbitrarily large public keys (#833).
 * Added Examples to Mdbook for Web Client (#850).
 * Added account code to `miden account --show` command (#835).
+* Changed exec's input file format to TOML instead of JSON (#870).
+* [BREAKING] Client's methods renamed after `PartialMmr` change to `PartialBlockchain` (#894).
 * Exposed new test utilities in the `testing` feature (#882).
 
 ## 0.8.2 (TBD)